refactor: 将VirtualFileSystemPlugin.ts拆分为多个模块便于维护
