+++ src/plugins/VirtualFileSystemPlugin/fileUtils.ts
@@ -1,0 +1,40 @@
+import * as fs from 'fs';
+import * as path from 'path';
+
+// 获取目录下所有的YAML文件
+export function getAllYamlFiles(dir: string): string[] {
+    const results: string[] = [];
+    const entries = fs.readdirSync(dir, { withFileTypes: true });
+    
+    for (const entry of entries) {
+        const fullPath = path.join(dir, entry.name);
+        if (entry.isDirectory()) {
+            results.push(...getAllYamlFiles(fullPath));
+        } else if (entry.isFile() && 
+                  (path.extname(entry.name) === '.yml' || 
+                   path.extname(entry.name) === '.yaml')) {
+            results.push(fullPath);
+        }
+    }
+    
+    return results;
+}
+
+// 获取目录下所有的Markdown文件
+export function getAllMarkdownFiles(dir: string): string[] {
+    const results: string[] = [];
+    const entries = fs.readdirSync(dir, { withFileTypes: true });
+    
+    for (const entry of entries) {
+        const fullPath = path.join(dir, entry.name);
+        if (entry.isDirectory()) {
+            results.push(...getAllMarkdownFiles(fullPath));
+        } else if (entry.isFile() && 
+                  (path.extname(entry.name) === '.md' || 
+                   path.extname(entry.name) === '.markdown')) {
+            results.push(fullPath);
+        }
+    }
+    
+    return results;
+}
